requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev.in
#
alabaster==0.7.12         # via sphinx
appdirs==1.4.3            # via virtualenv
attrs==19.3.0             # via pytest
babel==2.8.0              # via sphinx
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via pip-tools
coverage==5.0.4           # via -r dev.in, pytest-cov
distlib==0.3.0            # via virtualenv
docutils==0.16            # via sphinx
filelock==3.0.12          # via tox, virtualenv
idna==2.9                 # via requests
imagesize==1.2.0          # via sphinx
importlib-metadata==1.5.0  # via importlib-resources, pluggy, pytest, tox, virtualenv
importlib-resources==1.4.0  # via virtualenv
jinja2==2.11.1            # via numpydoc, sphinx
markupsafe==1.1.1         # via jinja2
more-itertools==8.2.0     # via pytest
numpydoc==0.9.2           # via -r dev.in
packaging==20.3           # via pytest, sphinx, tox
pip-tools==4.5.1          # via -r dev.in
pluggy==0.13.1            # via pytest, tox
py==1.8.1                 # via pytest, tox
pygments==2.6.1           # via sphinx
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1         # via -r dev.in
pytest==5.4.1             # via -r dev.in, pytest-cov
pytz==2019.3              # via babel
requests==2.23.0          # via sphinx
six==1.14.0               # via packaging, pip-tools, tox, virtualenv
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.4.3   # via -r dev.in
sphinx==2.4.4             # via -r dev.in, numpydoc, sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
toml==0.10.0              # via tox
tox==3.14.5               # via -r dev.in
urllib3==1.25.8           # via requests
virtualenv==20.0.13       # via tox
wcwidth==0.1.8            # via pytest
zipp==3.1.0               # via importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
